refactored and tidy the sudoku into new test pkg
