+++ test/sandbox/sudoku/sudoku_data_test.go
@@ -1,0 +1,35 @@
+package sudoku_test
+
+var datum_2 = []byte(`
+	{
+		"name": "sudoku.com some date",
+		"grid": [
+			3,0,0,1,0,8,0,6,0,
+			0,2,6,0,4,0,8,0,0,
+			0,0,1,0,0,0,4,0,0,
+			5,0,8,0,0,7,0,0,1,
+			0,0,0,0,9,5,7,0,0,
+			0,7,9,2,3,0,0,0,0,
+			9,0,0,0,0,0,5,0,6,
+			2,0,4,0,0,0,1,0,0,
+			6,0,0,5,0,0,0,2,0
+		]
+	},
+`)
+
+var datum_3 = []byte(`
+	{
+		"name": "sudoku.com/medium jul 25 2022",
+		"grid": [ 
+			0,0,5,0,0,6,8,0,7,
+			0,0,6,8,0,1,0,9,0,
+			0,9,1,4,0,0,0,0,5,
+			0,0,9,6,0,5,0,0,0,
+			0,0,0,0,0,4,0,0,0,
+			0,0,8,0,0,0,7,4,0,
+			9,0,0,0,0,0,0,0,0,
+			0,0,7,0,0,2,0,5,1,
+			2,0,4,1,9,0,6,3,0
+		]
+	}
+`)
